# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Airwave broadcast API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Audio source — "tab" for tab/screen audio, "mic" for microphone
 */
export type SourceType = (typeof SourceType)[keyof typeof SourceType];

export const SourceType = {
  tab: "tab",
  mic: "mic",
} as const;

export interface Room {
  /** Short room code used in the share URL */
  code: string;
  title: string;
  hostName: string;
  sourceType: SourceType;
  /** True when the host is currently connected via WebRTC */
  isLive: boolean;
  /** Current connected listeners */
  listenerCount: number;
  createdAt: string;
  endedAt?: string | null;
}

export interface CreateRoomRequest {
  /**
   * @minLength 1
   * @maxLength 80
   */
  title: string;
  /**
   * @minLength 1
   * @maxLength 40
   */
  hostName: string;
  sourceType: SourceType;
}

export interface CreateRoomResponse {
  room: Room;
  /** Secret token the host stores locally to authenticate signaling and end-broadcast actions */
  hostToken: string;
}

export interface EndRoomRequest {
  hostToken: string;
}

export interface PlatformStats {
  totalBroadcasts: number;
  liveNow: number;
  listenersConnected: number;
}

export interface ErrorResponse {
  error: string;
}