# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Airwave broadcast API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: rooms
    description: Broadcast room operations
  - name: stats
    description: Aggregate stats
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /rooms:
    get:
      operationId: listRooms
      tags: [rooms]
      summary: List recent broadcast rooms
      description: Returns recent broadcast rooms (live first, then most recently created), capped at 30.
      responses:
        "200":
          description: List of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create a new broadcast room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoomRequest"
      responses:
        "201":
          description: Room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateRoomResponse"

  /rooms/{code}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Get a room by short code
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /rooms/{code}/end:
    post:
      operationId: endRoom
      tags: [rooms]
      summary: End a broadcast (host only)
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EndRoomRequest"
      responses:
        "200":
          description: Room ended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "401":
          description: Invalid host token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stats:
    get:
      operationId: getStats
      tags: [stats]
      summary: Platform stats
      description: Lightweight aggregate stats — total broadcasts ever, live right now, listeners connected.
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    SourceType:
      type: string
      enum: [tab, mic]
      description: Audio source — "tab" for tab/screen audio, "mic" for microphone

    Room:
      type: object
      properties:
        code:
          type: string
          description: Short room code used in the share URL
        title:
          type: string
        hostName:
          type: string
        sourceType:
          $ref: "#/components/schemas/SourceType"
        isLive:
          type: boolean
          description: True when the host is currently connected via WebRTC
        listenerCount:
          type: integer
          description: Current connected listeners
        createdAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - code
        - title
        - hostName
        - sourceType
        - isLive
        - listenerCount
        - createdAt

    CreateRoomRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 80
        hostName:
          type: string
          minLength: 1
          maxLength: 40
        sourceType:
          $ref: "#/components/schemas/SourceType"
      required:
        - title
        - hostName
        - sourceType

    CreateRoomResponse:
      type: object
      properties:
        room:
          $ref: "#/components/schemas/Room"
        hostToken:
          type: string
          description: Secret token the host stores locally to authenticate signaling and end-broadcast actions
      required:
        - room
        - hostToken

    EndRoomRequest:
      type: object
      properties:
        hostToken:
          type: string
      required:
        - hostToken

    PlatformStats:
      type: object
      properties:
        totalBroadcasts:
          type: integer
        liveNow:
          type: integer
        listenersConnected:
          type: integer
      required:
        - totalBroadcasts
        - liveNow
        - listenersConnected

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error