# lib/api-zod/src/generated/types/createRoomRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Airwave broadcast API
 * OpenAPI spec version: 0.1.0
 */
import type { SourceType } from "./sourceType";

export interface CreateRoomRequest {
  /**
   * @minLength 1
   * @maxLength 80
   */
  title: string;
  /**
   * @minLength 1
   * @maxLength 40
   */
  hostName: string;
  sourceType: SourceType;
}