# lib/api-zod/src/generated/types/createRoomResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Airwave broadcast API
 * OpenAPI spec version: 0.1.0
 */
import type { Room } from "./room";

export interface CreateRoomResponse {
  room: Room;
  /** Secret token the host stores locally to authenticate signaling and end-broadcast actions */
  hostToken: string;
}