# lib/api-zod/src/generated/types/platformStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Airwave broadcast API
 * OpenAPI spec version: 0.1.0
 */

export interface PlatformStats {
  totalBroadcasts: number;
  liveNow: number;
  listenersConnected: number;
}