# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Airwave broadcast API
 * OpenAPI spec version: 0.1.0
 */
import type { SourceType } from "./sourceType";

export interface Room {
  /** Short room code used in the share URL */
  code: string;
  title: string;
  hostName: string;
  sourceType: SourceType;
  /** True when the host is currently connected via WebRTC */
  isLive: boolean;
  /** Current connected listeners */
  listenerCount: number;
  createdAt: Date;
  endedAt?: Date | null;
}