# lib/api-zod/src/generated/types/sourceType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Airwave broadcast API
 * OpenAPI spec version: 0.1.0
 */

/**
 * Audio source — "tab" for tab/screen audio, "mic" for microphone
 */
export type SourceType = (typeof SourceType)[keyof typeof SourceType];

export const SourceType = {
  tab: "tab",
  mic: "mic",
} as const;